fix(schema): distinguish Float32 and Float64 in migration generation
Previously, both Float32 and Float64 model columns mapped to the :float
symbol, causing db:generate to produce incorrect SQL. Float64 columns
would generate REAL (Float32) instead of DOUBLE PRECISION.

Changes:
- Float32 now maps to :float32 → REAL
- Float64 now maps to :float64 → DOUBLE PRECISION (PostgreSQL)

This ensures generated migrations match the precision of model types,
preventing Float32/Float64 type mismatches between models and database.

+++ src/ralph/schema/model_schema.cr
@@ -14,8 +14,8 @@ module Ralph
         "Int16"     => :smallint,
         "String"    => :string,
         "Bool"      => :boolean,
-        "Float64"   => :float,
-        "Float32"   => :float,
+        "Float64"   => :float64, # DOUBLE PRECISION
+        "Float32"   => :float32, # REAL
         "Time"      => :timestamp,
         "JSON::Any" => :jsonb,
         "Bytes"     => :binary,
